aider: Refactored ThemeController to load template based on active theme.
# Aider chat conversation:

USER: let's refactor ThemeController to load template based on active theme

ASSISTANT: To load a template based on the active theme, we need to modify the `listThemes` method in the `ThemeController` to pass the active theme to the template. We also need to modify the `themes.html.twig` template to use the active theme.

Here are the changes:

1. Modify `src/Controller/ThemeController.php`:

```php
src/Controller/ThemeController.php
<<<<<<< HEAD
    #[Route("/themes", name: "themes")]
    public function listThemes(): Response
    {
        $themes = $this->themeManager->findAllThemes();

        return $this->render('themes.html.twig', ['themes' => $themes]);
    }
=======
    #[Route("/themes", name: "themes")]
    public function listThemes(): Response
    {
        $themes = $this->themeManager->findAllThemes();
        $activeTheme = $this->themeManager->findThemeById($this->themeManager->getActiveThemeId());

        return $this->render('themes.html.twig', ['themes' => $themes, 'activeTheme' => $activeTheme]);
    }
>>>>>>> updated
```

2. Modify `templates/themes.html.twig`:

```twig
templates/themes.html.twig
<<<<<<< HEAD
{% block body %}
<h1>Available Themes</h1>
=======
{% block body %}
<h1>Available Themes (Active: {{ activeTheme.name }})</h1>
>>>>>>> updated
```

Please note that the `getActiveThemeId` method is not currently present in the `ThemeManager` class. You need to implement this method in the `ThemeManager` class. The method should return the ID of the currently active theme.