aider: Added `getActiveThemeId` method to `ThemeManager` to retrieve the ID of the currently active theme.
# Aider chat conversation:

USER: let's also update `ThemeManager`

ASSISTANT: To update the `ThemeManager` to include a method for getting the active theme, we need to add a `getActiveThemeId` method. This method will return the ID of the currently active theme.

Here are the changes:

1. Modify `src/Service/ThemeManager.php`:

```php
src/Service/ThemeManager.php
<<<<<<< HEAD
    public function deactivateTheme(int $id): void
    {
        $theme = $this->findThemeById($id);
        if ($theme) {
            $theme->isActive = false;
            $this->themeRepository->save($theme);
        }
    }
=======
    public function deactivateTheme(int $id): void
    {
        $theme = $this->findThemeById($id);
        if ($theme) {
            $theme->isActive = false;
            $this->themeRepository->save($theme);
        }
    }

    public function getActiveThemeId(): ?int
    {
        $themes = $this->findAllThemes();
        foreach ($themes as $theme) {
            if ($theme->isActive) {
                return $theme->id;
            }
        }

        return null;
    }
>>>>>>> updated
```

This new method `getActiveThemeId` will loop through all the themes and return the ID of the theme that is currently active. If no theme is active, it will return `null`.