aider: The ThemeRuntimeConfigurator now registers the themes folder in the Twig path loader.
# Aider chat conversation:

USER: it seems that ThemeRuntimeConfigurator doesn't register the themes folder in twig path loader, we are getting the following error Unable to find template "zenchron/nexus-theme/templates/admin/theme_manager/themes.html.twig" (looked into: /app/templates, /app/vendor/symfony/twig-bridge/Resources/views/Form).

ASSISTANT: The error message indicates that the Twig template engine is unable to find the specified template file. This is likely because the path to the themes directory is not correctly registered in the Twig environment.

To fix this, we need to modify the `configure` method in the `ThemeRuntimeConfigurator` service to correctly add the themes directory to the Twig loader paths.

Here is the change we need to make:

```php
src/Service/ThemeRuntimeConfigurator.php
<<<<<<< HEAD
    public function configure(): void
    {
        $activeThemeName = $this->themeManager->getActiveThemeName();
        if ($activeThemeName !== null) {
            $this->twig->getLoader()->addPath('themes/' . $activeThemeName);
        }
    }
=======
    public function configure(): void
    {
        $activeThemeName = $this->themeManager->getActiveThemeName();
        if ($activeThemeName !== null) {
            $this->twig->getLoader()->addPath('%kernel.project_dir%/themes/' . $activeThemeName, 'theme');
        }
    }
>>>>>>> updated
```

In the updated code, we use the `%kernel.project_dir%` parameter to get the absolute path to the project directory. This ensures that the path to the themes directory is correctly resolved regardless of where the application is run from. We also add a namespace 'theme' to the path, so we can refer to the templates in a more straightforward way, like `@theme/admin/theme_manager/themes.html.twig`.