aider: Updated `ThemeRuntimeConfigurator` to inject `kernel.project_dir` parameter and added service definition in `services.yaml`.
# Aider chat conversation:

USER: We have another error now that %kernel.project_dir% is not availabie in `themeRuntimeConfigurator`

ASSISTANT: The `%kernel.project_dir%` parameter is a service container parameter that holds the absolute path to your project directory. It's available in the service container, but not directly in your services unless you inject it.

To fix this, we need to modify the `ThemeRuntimeConfigurator` service to inject the `kernel.project_dir` parameter.

Here is the change we need to make:

```php
src/Service/ThemeRuntimeConfigurator.php
<<<<<<< HEAD
class ThemeRuntimeConfigurator
{
    private Environment $twig;
    private ThemeManager $themeManager;

    public function __construct(Environment $twig, ThemeManager $themeManager)
    {
        $this->twig = $twig;
        $this->themeManager = $themeManager;
    }

    public function configure(): void
    {
        $activeThemeName = $this->themeManager->getActiveThemeName();
        if ($activeThemeName !== null) {
            $this->twig->getLoader()->addPath('%kernel.project_dir%/themes/' . $activeThemeName, 'theme');
        }
    }
}
=======
class ThemeRuntimeConfigurator
{
    private Environment $twig;
    private ThemeManager $themeManager;
    private string $projectDir;

    public function __construct(Environment $twig, ThemeManager $themeManager, string $projectDir)
    {
        $this->twig = $twig;
        $this->themeManager = $themeManager;
        $this->projectDir = $projectDir;
    }

    public function configure(): void
    {
        $activeThemeName = $this->themeManager->getActiveThemeName();
        if ($activeThemeName !== null) {
            $this->twig->getLoader()->addPath($this->projectDir . '/themes/' . $activeThemeName, 'theme');
        }
    }
}
>>>>>>> updated
```

In the updated code, we inject the `kernel.project_dir` parameter into the `ThemeRuntimeConfigurator` service and store it in the `$projectDir` property. We then use this property when adding the path to the Twig loader.

Please note that you also need to update the service definition in your Symfony service configuration file (usually `services.yaml`) to pass the `kernel.project_dir` parameter to the `ThemeRuntimeConfigurator` service. Here is how you can do it:

```yaml
services.yaml
<<<<<<< HEAD
    App\Service\ThemeRuntimeConfigurator:
        arguments:
            $twig: '@twig'
            $themeManager: '@App\Service\ThemeManager'
=======
    App\Service\ThemeRuntimeConfigurator:
        arguments:
            $twig: '@twig'
            $themeManager: '@App\Service\ThemeManager'
            $projectDir: '%kernel.project_dir%'
>>>>>>> updated
```

In the updated service definition, we add a new argument `$projectDir` and set its value to the `kernel.project_dir` parameter.

USER: InvalidEditBlock: edit failed!

services.yaml does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in services.yaml with nothing missing!

services.yaml does not contain these 4 exact lines in a row:
```
    App\Service\ThemeRuntimeConfigurator:
        arguments:
            $twig: '@twig'
            $themeManager: '@App\Service\ThemeManager'
```

ASSISTANT: I apologize for the confusion. It seems that the `services.yaml` file does not contain the exact configuration for `ThemeRuntimeConfigurator` as I assumed.

In this case, you need to add the `ThemeRuntimeConfigurator` service definition to your `services.yaml` file. Here is how you can do it:

```yaml
services.yaml
<<<<<<< HEAD

=======
    App\Service\ThemeRuntimeConfigurator:
        arguments:
            $twig: '@twig'
            $themeManager: '@App\Service\ThemeManager'
            $projectDir: '%kernel.project_dir%'
>>>>>>> updated
```

In the updated service definition, we add a new service `App\Service\ThemeRuntimeConfigurator` and set its arguments `$twig`, `$themeManager`, and `$projectDir`. The `$twig` and `$themeManager` arguments are set to the `@twig` and `@App\Service\ThemeManager` services respectively, and the `$projectDir` argument is set to the `kernel.project_dir` parameter.