aider: Modified `templates/table_row.html.twig` to correctly check if packages are installed by passing the `installedPackages` array to the included `table_row.html.twig` file and checking if the package is in the array.
# Aider chat conversation:

USER: message Installed / Not Installed does not appear in `templates/table_row.html.twig`

ASSISTANT: The issue seems to be that the installed packages are not being checked correctly in the `templates/table_row.html.twig` file. The condition `{% elseif key in ['Require', 'Require Dev'] %}` is not being met because the `key` variable is not exactly 'Require' or 'Require Dev'. Instead, it's the name of the package.

We need to modify the `templates/table_row.html.twig` file to check if the package is in the `installedPackages` array for each package in `Require` and `Require Dev`.

Here is the proposed change:

```twig
templates/table_row.html.twig
<<<<<<< HEAD
<tr>
    <td>{{ key }}</td>
    <td>
        {% if value is iterable %}
            <ul>
                {% for subKey, subValue in value %}
                {% include 'table_row.html.twig' with {'key': subKey, 'value': subValue} %}
                {% endfor %}
            </ul>
        {% elseif key in ['Require', 'Require Dev'] %}
            {{ value }} ({{ value in installedPackages ? 'Installed' : 'Not Installed' }})
        {% else %}
            {{ value }}
        {% endif %}
    </td>
</tr>
=======
<tr>
    <td>{{ key }}</td>
    <td>
        {% if value is iterable %}
            <ul>
                {% for subKey, subValue in value %}
                {% include 'table_row.html.twig' with {'key': subKey, 'value': subValue, 'installedPackages': installedPackages} %}
                {% endfor %}
            </ul>
        {% else %}
            {{ value }} ({{ key in installedPackages ? 'Installed' : 'Not Installed' }})
        {% endif %}
    </td>
</tr>
>>>>>>> updated
```

In the updated code, we pass the `installedPackages` array to the included `table_row.html.twig` file when the `value` is iterable. Then, we check if the `key` (which is the package name) is in the `installedPackages` array. If it is, we display 'Installed', otherwise we display 'Not Installed'.