aider: Pass the `installedPackages` array to the template in `ComposerJsonVisualizerController.php`.
# Aider chat conversation:

USER: variabile installedPackages is missing / undefined probably it needs to be defined in `src/Controller/ComposerJsonVisualizerController.php`

ASSISTANT: You're correct. The `installedPackages` variable is not being passed to the template in the `src/Controller/ComposerJsonVisualizerController.php` file. We need to modify the `__invoke` method in this file to pass the `installedPackages` array to the template.

Here is the proposed change:

```php
src/Controller/ComposerJsonVisualizerController.php
<<<<<<< HEAD
    #[Route("/visualize", name: "composer_json_visualize")]
    public function __invoke(): Response
    {
        $composerJsonData = $this->composerJsonVisualizer->visualize();

        return $this->render('composer_json_visualizer.html.twig', ['composerJsonData' => $composerJsonData]);
    }
=======
    #[Route("/visualize", name: "composer_json_visualize")]
    public function __invoke(): Response
    {
        $composerJsonData = $this->composerJsonVisualizer->visualize();

        return $this->render('composer_json_visualizer.html.twig', [
            'composerJsonData' => $composerJsonData,
            'installedPackages' => $composerJsonData->installedPackages
        ]);
    }
>>>>>>> updated
```

In the updated code, we pass the `installedPackages` array to the template by adding it to the array of variables passed to the `render` method. The `installedPackages` array is accessed from the `composerJsonData` object.