aider: Fix bug in `templates/table_row.html.twig` to display 'Installed' or 'Not Installed' only for 'Require' and 'Require Dev' keys.
# Aider chat conversation:

USER: let's fix a bug in `templates/table_row.html.twig` and before displaying Installed or Not Installed to check if key is equal with require or require-dev

ASSISTANT: To fix the bug in `templates/table_row.html.twig`, we need to add a condition to check if the key is 'Require' or 'Require Dev' before displaying 'Installed' or 'Not Installed'.

Here is the proposed change:

```twig
templates/table_row.html.twig
<<<<<<< HEAD
    <td>
        {% if value is iterable %}
            <ul>
                {% for subKey, subValue in value %}
                {% include 'table_row.html.twig' with {'key': subKey, 'value': subValue, 'installedPackages': installedPackages} %}
                {% endfor %}
            </ul>
        {% else %}
            {{ value }} ({{ key in installedPackages ? 'Installed' : 'Not Installed' }})
        {% endif %}
    </td>
=======
    <td>
        {% if value is iterable %}
            <ul>
                {% for subKey, subValue in value %}
                {% include 'table_row.html.twig' with {'key': subKey, 'value': subValue, 'installedPackages': installedPackages} %}
                {% endfor %}
            </ul>
        {% else %}
            {{ value }} {% if key == 'Require' or key == 'Require Dev' %} ({{ key in installedPackages ? 'Installed' : 'Not Installed' }}) {% endif %}
        {% endif %}
    </td>
>>>>>>> updated
```

This change will ensure that the 'Installed' or 'Not Installed' status is only displayed for the 'Require' and 'Require Dev' keys.