aider: Added service definition for `InstallAssetsController` in `config/services.yaml` with dependencies `ThemeManager` and `AssetManager`.
# Aider chat conversation:

USER: let's fix a bug and create a service for `InstallAssetsController` in `config/services.yaml`

ASSISTANT: To create a service for `InstallAssetsController` in `config/services.yaml`, we need to add a new service definition for `InstallAssetsController` under the `services` section. This service will have two arguments, `ThemeManager` and `AssetManager`, which are the dependencies required by the `InstallAssetsController`.

Here is the *edit block* for this change:

```yaml
config/services.yaml
<<<<<<< HEAD
    App\Service\ComposerJsonVisualizer:
        public: true
        arguments:
            $projectDir: '%kernel.project_dir%'
=======
    App\Service\ComposerJsonVisualizer:
        public: true
        arguments:
            $projectDir: '%kernel.project_dir%'

    App\Controller\InstallAssetsController:
        arguments:
            $themeManager: '@App\Service\ThemeManager'
            $assetManager: '@App\Service\AssetManager'
>>>>>>> updated
```

This change will register `InstallAssetsController` as a service and inject the `ThemeManager` and `AssetManager` services into it.