aider: Added a service definition for `ThemeManager` in `config/services.yaml` with dependencies `ThemeRepository` and `ThemeFilesystemService`.
# Aider chat conversation:

USER: let's refactor `services.yaml` and add a service for `src/Service/ThemeManager.php`

ASSISTANT: To create a service for `src/Service/ThemeManager.php` in `config/services.yaml`, we need to add a new service definition for `ThemeManager` under the `services` section. This service will have two arguments, `ThemeRepository` and `ThemeFilesystemService`, which are the dependencies required by the `ThemeManager`.

Here is the *edit block* for this change:

```yaml
config/services.yaml
<<<<<<< HEAD
    App\Service\ComposerJsonVisualizer:
        public: true
        arguments:
            $projectDir: '%kernel.project_dir%'

    App\Controller\InstallAssetsController:
        arguments:
            $themeManager: '@App\Service\ThemeManager'
            $assetManager: '@App\Service\AssetManager'
=======
    App\Service\ComposerJsonVisualizer:
        public: true
        arguments:
            $projectDir: '%kernel.project_dir%'

    App\Controller\InstallAssetsController:
        arguments:
            $themeManager: '@App\Service\ThemeManager'
            $assetManager: '@App\Service\AssetManager'

    App\Service\ThemeManager:
        arguments:
            $themeRepository: '@App\Repository\ThemeRepository'
            $themeFilesystemService: '@App\Service\ThemeFilesystemService'
>>>>>>> updated
```

This change will register `ThemeManager` as a service and inject the `ThemeRepository` and `ThemeFilesystemService` services into it.